requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
